In PowerPC as, keep a list of hi16 and ha16 relocations.
Use it to pair hi16 and ha16 relocations with lo16 relocations even if
there are other instructions between them.  For now, the linker still
requires the lo16 in the next instruction.

Code my own linked lists and a bitset because I don't have convenient
libraries for them.  My bitset is an array of 16-bit values, like the
bitsets in ncgg (util/ncgg/set.h), but I also use realloc() to grow
the bitset.